update Flutter version from 3.7.12 to 3.16.0

+++ lib/src/views/explore/explore.dart
@@ -105,7 +105,7 @@ class _SearchPageState extends State<SearchPage> {
                         activeColor: Colors.red,
                         color: Colors.white),
                     dotsCount: imageSample.length,
-                    position: currentIndex.toDouble(),
+                    position: currentIndex,
                   ),
                 ),
               ],
